[AAE-8748] - Add abstraction layer for JWT token helper service
