+++ docs/core/services/jwt-helper.service.md
@@ -2,7 +2,7 @@
 Title: JWT helper service
 Added: v3.0.0
 Status: Active
-Last reviewed: 2019-01-09
+Last reviewed: 2022-05-11
 ---
 
 # [JWT helper service](../../../lib/core/services/jwt-helper.service.ts "Defined in jwt-helper.service.ts")
@@ -13,36 +13,13 @@ Decodes a JSON Web Token (JWT) to a JavaScript object.
 
 ### Methods
 
--   **decodeToken**(token: `any`): `Object`<br/>
-    Decodes a JSON web token into a JS object.
-    -   _token:_ `any`  - Token in encoded form
-    -   **Returns** `Object` - Decoded token data object
--   **getAccessToken**(): `string`<br/>
-    Gets access token
-    -   **Returns** `string` - access token
 -   **getClientRoles**(clientName: `string`): `string[]`<br/>
     Gets Client roles.
     -   _clientName:_ `string`  - 
     -   **Returns** `string[]` - Array of client roles
--   **getIdToken**(): `string`<br/>
-    Gets id token
-    -   **Returns** `string` - id token
 -   **getRealmRoles**(): `string[]`<br/>
     Gets realm roles.
     -   **Returns** `string[]` - Array of realm roles
--   **getValueFromLocalAccessToken**(key: `string`)<br/>
-    Gets a named value from the user access token.
-    -   _key:_ `string`  - Key name of the field to retrieve
--   **getValueFromLocalIdToken**(key: `string`)<br/>
-    Gets a named value from the user id token.
-    -   _key:_ `string`  - Key name of the field to retrieve
--   **getValueFromLocalToken**(key: `string`)<br/>
-    Gets a named value from the user access or id token.
-    -   _key:_ `string`  - Key name of the field to retrieve
--   **getValueFromToken**(token: `string`, key: `string`)<br/>
-    Gets a named value from the user access token.
-    -   _token:_ `string`  - 
-    -   _key:_ `string`  - Key name of the field to retrieve
 -   **hasClientRole**(clientName: `string`, role: `string`): `boolean`<br/>
     Checks for client role.
     -   _clientName:_ `string`  - Targeted client name
@@ -64,9 +41,9 @@ Decodes a JSON Web Token (JWT) to a JavaScript object.
 
 ## Details
 
-JWT is a standard for sending data securely that ADF uses during the
-OAuth2 authentication procedure. See the [JWT website](https://jwt.io/)
-for full details of the standard and its uses.
+This service extends the [Base JWT helper service](base-jwt-helper.service.md) 
+which is responsible for reading the token value from the local storage and decoding it. 
+It can be used to check the realm and resource access defined in a JWT token.
 
 ## See also
