+++ lib/core/mock/public-api.ts
@@ -36,7 +36,7 @@ export * from './form/form-definition-visibility.mock';
 export * from './form/start-form.component.mock';
 export * from './form/form.service.mock';
 export * from './form/widget-visibility.service.mock';
-export * from './jwt-helper.service.spec';
+export * from './jwt-token.mock';
 export * from './identity-group.mock';
 export * from './identity-user.mock';
 export * from './identity-group.service.mock';
