+++ lib/core/services/identity-user.service.spec.ts
@@ -29,7 +29,7 @@ import { IdentityUserService } from './identity-user.service';
 import { JwtHelperService } from './jwt-helper.service';
 import { setupTestBed } from '../testing/setup-test-bed';
 import { AlfrescoApiService } from './alfresco-api.service';
-import { mockToken } from '../mock/jwt-helper.service.spec';
+import { mockToken } from '../mock/jwt-token.mock';
 import { IdentityRoleModel } from '../models/identity-role.model';
 import { CoreTestingModule } from '../testing/core.testing.module';
 import { TranslateModule } from '@ngx-translate/core';
