+++ lib/core/services/jwt-helper.service.ts
@@ -16,131 +16,15 @@
  */
 
 import { Injectable } from '@angular/core';
-import { StorageService } from './storage.service';
+import { JwtHelper } from './authentication/jwt-helper.interface';
+import { BaseJwtHelperService } from './authentication/base-jwt-helper.service';
 
 @Injectable({
     providedIn: 'root'
 })
-export class JwtHelperService {
-
-    static USER_NAME = 'name';
-    static FAMILY_NAME = 'family_name';
-    static GIVEN_NAME = 'given_name';
-    static USER_EMAIL = 'email';
-    static USER_ACCESS_TOKEN = 'access_token';
-    static USER_ID_TOKEN = 'id_token';
+export class JwtHelperService extends BaseJwtHelperService implements JwtHelper {
     static REALM_ACCESS = 'realm_access';
     static RESOURCE_ACCESS = 'resource_access';
-    static USER_PREFERRED_USERNAME = 'preferred_username';
-
-    constructor(private storageService: StorageService) {
-    }
-
-    /**
-     * Decodes a JSON web token into a JS object.
-     *
-     * @param token Token in encoded form
-     * @returns Decoded token data object
-     */
-    decodeToken(token): any {
-        const parts = token.split('.');
-
-        if (parts.length !== 3) {
-            throw new Error('JWT must have 3 parts');
-        }
-
-        const decoded = this.urlBase64Decode(parts[1]);
-        if (!decoded) {
-            throw new Error('Cannot decode the token');
-        }
-
-        return JSON.parse(decoded);
-    }
-
-    private urlBase64Decode(token): string {
-        let output = token.replace(/-/g, '+').replace(/_/g, '/');
-        switch (output.length % 4) {
-            case 0: {
-                break;
-            }
-            case 2: {
-                output += '==';
-                break;
-            }
-            case 3: {
-                output += '=';
-                break;
-            }
-            default: {
-                throw new Error('Illegal base64url string!');
-            }
-        }
-        return decodeURIComponent(escape(window.atob(output)));
-    }
-
-    /**
-     * Gets a named value from the user access or id token.
-     *
-     * @param key Key name of the field to retrieve
-     * @returns Value from the token
-     */
-     getValueFromLocalToken<T>(key: string): T {
-        return this.getValueFromToken(this.getAccessToken(), key) || this.getValueFromToken(this.getIdToken(), key);
-    }
-
-    /**
-     * Gets a named value from the user access token.
-     *
-     * @param key Key name of the field to retrieve
-     * @returns Value from the token
-     */
-    getValueFromLocalAccessToken<T>(key: string): T {
-        return this.getValueFromToken(this.getAccessToken(), key);
-    }
-
-    /**
-     * Gets access token
-     *
-     * @returns access token
-     */
-    getAccessToken(): string {
-        return this.storageService.getItem(JwtHelperService.USER_ACCESS_TOKEN);
-    }
-
-    /**
-     * Gets a named value from the user id token.
-     *
-     * @param key Key name of the field to retrieve
-     * @returns Value from the token
-     */
-     getValueFromLocalIdToken<T>(key: string): T {
-        return this.getValueFromToken(this.getIdToken(), key);
-    }
-
-    /**
-     * Gets id token
-     *
-     * @returns id token
-     */
-     getIdToken(): string {
-        return this.storageService.getItem(JwtHelperService.USER_ID_TOKEN);
-    }
-
-    /**
-     * Gets a named value from the user access token.
-     *
-     * @param accessToken your SSO access token where the value is encode
-     * @param key Key name of the field to retrieve
-     * @returns Value from the token
-     */
-    getValueFromToken<T>(token: string, key: string): T {
-        let value;
-        if (token) {
-            const tokenPayload = this.decodeToken(token);
-            value = tokenPayload[key];
-        }
-        return value;
-    }
 
     /**
      * Gets realm roles.
